Bug 1674777 part 9 - Change GetArrayBufferByteLength return type from uint32_t to size_t. r=sfink
Note that previous patches already fixed various other callers.

Depends on D103760

Differential Revision: https://phabricator.services.mozilla.com/D103761

UltraBlame original commit: 42239f7a0ddf8b6024b464a847d87f9e0a56d42a

+++ js/public/ArrayBuffer.h
@@ -207,7 +207,7 @@ extern JS_PUBLIC_API JSObject* GetObjectAsArrayBuffer(JSObject* obj,
 
 
 
-extern JS_PUBLIC_API uint32_t GetArrayBufferByteLength(JSObject* obj);
+extern JS_PUBLIC_API size_t GetArrayBufferByteLength(JSObject* obj);
